+++ Mod Info/changelog.md
@@ -1,2 +1,3 @@
 1/26/21
+- Create `DebugManager` to output debug info to `output.txt` file. Output a test "GAME STARTED" in `LudositySplash.Awake()`
 - Create project. Realized I took a wrong turn in life that has led me to rewrite 18k+ lines of code for a mod that a handful of people would use :)